+++ src/backend/mod.rs
@@ -69,7 +69,7 @@ pub trait DBColumnRefBatch<'c>: DBColumn {
     type Ref: AsRef<[u8]> + 'c + std::ops::Deref<Target = [u8]> + Send + Sync;
 
     /// Get a value by key in batch.
-    fn get_multi_ref<I>(&self, keys: I) -> Result<Vec<Option<Self::Ref>>>
+    fn get_multi_ref<I>(&'c self, keys: I) -> Result<Vec<Option<Self::Ref>>>
     where
         I: IntoIterator,
         I::Item: AsRef<[u8]>;
@@ -80,7 +80,7 @@ pub trait DBColumnTransaction<'c>: DBColumn {
     type Txn: DBTransaction;
 
     /// Start a transaction.
-    fn transaction(&self) -> Result<Self::Txn>;
+    fn transaction(&'c self) -> Result<Self::Txn>;
 }
 
 /// Database transaction trait.