+++ src/db/database.rs
@@ -9,7 +9,7 @@ use std::sync::Arc;
 /// A collection of key-value pairs
 /// Can be cloned but not shared across threads.
 pub struct Database<K, V, D: DatabaseBackend>(Arc<DatabaseInner<K, V, D>>);
-impl<'a, K, V, D: DatabaseBackend> Database<K, V, D> {
+impl<K, V, D: DatabaseBackend> Database<K, V, D> {
     pub(crate) fn new(env: Env<D>, name: &str) -> Result<Self> {
         let column = D::create_or_open(env, name)?;
         Ok(Database(Arc::new(DatabaseInner {
@@ -119,7 +119,7 @@ where
 }
 
 #[inherent]
-impl<'a, Key, Val, D, C> crate::DBCommonDelete for Database<Key, Val, D>
+impl<Key, Val, D, C> crate::DBCommonDelete for Database<Key, Val, D>
 where
     C: DBColumnDelete,
     D: DatabaseBackend<Column = C>,
@@ -167,14 +167,17 @@ where
 #[inherent]
 impl<'a, Key, Val, D, C> crate::DBCommonRefBatch<'a, Key, Val, C::Ref> for Database<Key, Val, D>
 where
-    C: DBColumnRefBatch<'a>,
+    C: DBColumnRefBatch<'a> + 'a,
     D: DatabaseBackend<Column = C>,
 {
     /// Get the serialized `val` from the database by `key`.
     ///
     /// See [`get_multi_ref`](crate::DBCommonRefBatch::get_multi_ref) for more information.
-    #[allow(clippy::type_complexity)] // trait associated types are not stable yet
-    pub fn get_multi_ref<'k, I>(&self, keys: I) -> Result<Vec<Option<RefValue<C::Ref, Val::DItem>>>>
+    #[allow(clippy::type_complexity)] // this isn't that complex
+    pub fn get_multi_ref<'k, I>(
+        &'a self,
+        keys: I,
+    ) -> Result<Vec<Option<RefValue<C::Ref, Val::DItem>>>>
     where
         Key: BytesEncode<'k>,
         I: IntoIterator<Item = &'k <Key>::EItem>,
@@ -199,11 +202,11 @@ where
 // Databases that support transactions
 impl<'a, Key, Val, D, C> Database<Key, Val, D>
 where
-    C: DBColumnTransaction<'a>,
+    C: DBColumnTransaction<'a> + 'a,
     D: DatabaseBackend<Column = C>,
 {
     /// Clear the database, removing all key-value pairs.
-    pub fn transaction(&self) -> Result<DatabaseTransaction<'a, Key, Val, C>> {
+    pub fn transaction(&'a self) -> Result<DatabaseTransaction<'a, Key, Val, C>> {
         Ok(DatabaseTransaction {
             column: self.0.column.transaction()?,
             _phantom: PhantomData,
@@ -224,7 +227,7 @@ where
     }
 }
 
-impl<'a, K, V, D> Clone for Database<K, V, D>
+impl<K, V, D> Clone for Database<K, V, D>
 where
     D: DatabaseBackend,
 {
@@ -234,7 +237,7 @@ where
 }
 
 // Databases that support flushing
-impl<'a, K, V, D, C> Database<K, V, D>
+impl<K, V, D, C> Database<K, V, D>
 where
     C: DBColumn + Flushable,
     D: DatabaseBackend<Column = C>,